requirements.txt
netmiko==4.3.0
PyYAML==6.0.1
rich==13.7.1